+++ CHANGELOG.md
@@ -5,11 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.1.0] - 2023-13-19
 ### Fixed
 - Handle failure in purl creation gracefully (custom components will probably be skipped)
 - Support file with evidence but no copyright or license
 - Encoding fix for copyrights
 - Improved debugging
+### Added
 - LicenseRef values
 - Add package dependency support